refactor(ops): extract format/repr helpers and SemanticIndexOp
Two more chunks out of _core.py:

- _format.py (159 lines): _collect_chain, _format_op_summary,
  _format_root, _semantic_repr — the helpers that build the multi-line
  pipeline repr used by every Semantic*Op.__repr__. All Op classes are
  imported lazily inside each helper to dodge the circular dependency.

- _index_op.py (343 lines): SemanticIndexOp plus its 6 fragment
  builders (_get_field_type_str, _get_weight_expr,
  _build_string_index_fragment, _build_numeric_index_fragment,
  _resolve_selector, _get_fields_to_index). Self-contained except for
  lazy imports of _get_merged_fields/_to_untagged from _core.

_core.py is now 3,775 lines. All extractions are pure moves with
re-imports back so internal call sites are unchanged. 930 tests pass.
