+++ src/boring_semantic_layer/ops/_core.py
@@ -126,138 +126,7 @@ def _unwrap(wrapped: Any) -> Any:
     return wrapped.unwrap if isinstance(wrapped, _CallableWrapper) else wrapped
 
 
-def _collect_chain(op: Relation) -> list[Relation]:
-    """Walk op.source (or op.left for joins) back to root, return list from root to current."""
-    chain = [op]
-    current = op
-    while True:
-        if hasattr(current, "source") and current.source is not None:
-            chain.append(current.source)
-            current = current.source
-        elif hasattr(current, "left") and current.left is not None:
-            chain.append(current.left)
-            current = current.left
-        else:
-            break
-    chain.reverse()
-    return chain
-
-
-def _format_op_summary(op: Relation) -> str:
-    """Return a one-line summary string for a non-root semantic op."""
-    # Import here to avoid circular imports at module level
-    cls = type(op).__name__
-
-    if isinstance(op, SemanticFilterOp):
-        predicate = object.__getattribute__(op, "predicate")
-        pred_name = "<predicate>"
-        if hasattr(predicate, "__name__"):
-            pred_name = predicate.__name__
-        elif hasattr(predicate, "unwrap"):
-            unwrapped = predicate.unwrap
-            if hasattr(unwrapped, "__name__"):
-                pred_name = unwrapped.__name__
-        return f"Filter(\u03bb {pred_name})"
-
-    if isinstance(op, SemanticMutateOp):
-        post = object.__getattribute__(op, "post")
-        cols = list(post.keys())
-        return f"Mutate({', '.join(cols)})"
-
-    if isinstance(op, SemanticGroupByOp):
-        keys = object.__getattribute__(op, "keys")
-        return f"GroupBy({', '.join(keys)})"
-
-    if isinstance(op, SemanticAggregateOp):
-        aggs = object.__getattribute__(op, "aggs")
-        agg_names = list(aggs.keys())
-        return f"Aggregate({', '.join(agg_names)})"
-
-    if isinstance(op, SemanticOrderByOp):
-        keys = object.__getattribute__(op, "keys")
-        key_strs = [k if isinstance(k, str) else repr(k) for k in keys]
-        return f"OrderBy({', '.join(key_strs)})"
-
-    if isinstance(op, SemanticLimitOp):
-        return f"Limit({op.n})"
-
-    if isinstance(op, SemanticProjectOp):
-        fields = object.__getattribute__(op, "fields")
-        return f"Project({', '.join(fields)})"
-
-    if isinstance(op, SemanticUnnestOp):
-        column = object.__getattribute__(op, "column")
-        return f"Unnest({column})"
-
-    if isinstance(op, SemanticJoinOp):
-        how = object.__getattribute__(op, "how")
-        right = object.__getattribute__(op, "right")
-        right_name = ""
-        if isinstance(right, SemanticTableOp):
-            right_name = object.__getattribute__(right, "name") or ""
-        if not right_name:
-            # Try to find a root name from right side
-            roots = _find_all_root_models(right)
-            if roots:
-                right_name = object.__getattribute__(roots[0], "name") or ""
-        if right_name:
-            return f"Join({how}, right={right_name})"
-        return f"Join({how})"
-
-    if isinstance(op, SemanticIndexOp):
-        parts = []
-        selector = object.__getattribute__(op, "selector")
-        by = object.__getattribute__(op, "by")
-        sample = object.__getattribute__(op, "sample")
-        if selector is not None:
-            if isinstance(selector, tuple):
-                parts.append(", ".join(selector))
-            else:
-                parts.append(str(selector))
-        if by is not None:
-            parts.append(f"by={by}")
-        if sample is not None:
-            parts.append(f"sample={sample}")
-        return f"Index({', '.join(parts)})"
-
-    # Fallback for unknown op types
-    return cls.replace("Semantic", "").replace("Op", "")
-
-
-def _format_root(root_op: SemanticTableOp) -> str:
-    """Format a SemanticTableOp root using the fmt registry from format.py."""
-    from boring_semantic_layer.format import fmt
-
-    try:
-        return fmt(root_op)
-    except Exception:
-        # Fallback if format module isn't available
-        name = object.__getattribute__(root_op, "name")
-        return f"SemanticTable: {name}" if name else "SemanticTable"
-
-
-def _semantic_repr(op: Relation) -> str:
-    chain = _collect_chain(op)
-
-    # Find the root (first element should be a SemanticTableOp)
-    root = chain[0]
-    if isinstance(root, SemanticTableOp):
-        lines = [_format_root(root)]
-    else:
-        # Fallback: no SemanticTableOp root found
-        from ibis.expr.format import pretty
-
-        try:
-            return pretty(op)
-        except Exception:
-            return object.__repr__(op)
-
-    # Append pipeline steps
-    for step in chain[1:]:
-        if not isinstance(step, SemanticTableOp):
-            lines.append(f"-> {_format_op_summary(step)}")
-
-    return "\n".join(lines)
+from ._format import _collect_chain, _format_op_summary, _format_root, _semantic_repr  # noqa: E402, F401
 
 
 def _make_schema(fields_dict: dict[str, str]):
@@ -3866,295 +3735,15 @@ class SemanticLimitOp(Relation):
         return self.source.get_calculated_measures()
 
 
-def _get_field_type_str(field_type: Any) -> str:
-    return (
-        "string"
-        if field_type.is_string()
-        else "number"
-        if field_type.is_numeric()
-        else "date"
-        if field_type.is_temporal()
-        else str(field_type)
-    )
-
-
-def _get_weight_expr(
-    base_tbl: Any,
-    by_measure: str | None,
-    all_roots: list,
-    is_string: bool,
-) -> Any:
-    from .._xorq import api as xo
-
-    if not by_measure:
-        return xo._.count()
-
-    merged_measures = _get_merged_fields(all_roots, "measures")
-    return (
-        merged_measures[by_measure](base_tbl) if by_measure in merged_measures else xo._.count()
-    )
-
-
-def _build_string_index_fragment(
-    base_tbl: Any,
-    field_expr: Any,
-    field_name: str,
-    field_path: str,
-    type_str: str,
-    weight_expr: Any,
-) -> Any:
-    from .._xorq import api as xo
-
-    return (
-        base_tbl.group_by(field_expr.name("value"))
-        .aggregate(weight=weight_expr)
-        .select(
-            fieldName=xo.literal(field_name.split(".")[-1]),
-            fieldPath=xo.literal(field_path),
-            fieldType=xo.literal(type_str),
-            fieldValue=xo._["value"].cast("string"),
-            weight=xo._["weight"],
-        )
-    )
-
-
-def _build_numeric_index_fragment(
-    base_tbl: Any,
-    field_expr: Any,
-    field_name: str,
-    field_path: str,
-    type_str: str,
-    weight_expr: Any,
-) -> Any:
-    from .._xorq import api as xo
-
-    return (
-        base_tbl.select(field_expr.name("value"))
-        .filter(xo._["value"].notnull())
-        .aggregate(
-            min_val=xo._["value"].min(),
-            max_val=xo._["value"].max(),
-            weight=weight_expr,
-        )
-        .select(
-            fieldName=xo.literal(field_name.split(".")[-1]),
-            fieldPath=xo.literal(field_path),
-            fieldType=xo.literal(type_str),
-            fieldValue=(
-                xo._["min_val"].cast("string") + " to " + xo._["max_val"].cast("string")
-            ),
-            weight=xo._["weight"],
-        )
-    )
-
-
-def _resolve_selector(
-    selector: str | list[str] | Callable | None,
-    base_tbl: ir.Table,
-) -> tuple[str, ...]:
-    if selector is None:
-        return tuple(base_tbl.columns)
-    try:
-        selected = base_tbl.select(selector)
-        return tuple(selected.columns)
-    except Exception:
-        return []
-
-
-def _get_fields_to_index(
-    selector: str | list[str] | Callable | None,
-    merged_dimensions: dict,
-    base_tbl: ir.Table,
-) -> tuple[str, ...]:
-    if selector is None:
-        selector = s.all()
-
-    raw_fields = _resolve_selector(selector, base_tbl)
-
-    if not raw_fields:
-        result = list(merged_dimensions.keys())
-        result.extend(col for col in base_tbl.columns if col not in result)
-    else:
-        result = [col for col in raw_fields if col in merged_dimensions or col in base_tbl.columns]
-
-    return result
-
-
-class SemanticIndexOp(Relation):
-    source: Relation
-    selector: str | list[str] | tuple[str, ...] | Callable | None
-    by: str | None = None
-    sample: int | None = None
-
-    def __init__(
-        self,
-        source: Relation,
-        selector: str | list[str] | tuple[str, ...] | Callable | None = None,
-        by: str | None = None,
-        sample: int | None = None,
-    ) -> None:
-        # Validate sample parameter
-        if sample is not None and sample <= 0:
-            raise ValueError(f"sample must be positive, got {sample}")
-
-        # Validate 'by' measure exists if provided
-        if by is not None:
-            all_roots = _find_all_root_models(source)
-            if all_roots:
-                merged_measures = _get_merged_fields(all_roots, "measures")
-                if by not in merged_measures:
-                    available = list(merged_measures.keys())
-                    raise KeyError(
-                        f"Unknown measure '{by}' for weight calculation. "
-                        f"Available measures: {', '.join(available) or 'none'}",
-                    )
-
-        # Convert selector to tuple if it's a list (Ibis requires hashable types)
-        hashable_selector = tuple(selector) if isinstance(selector, list) else selector
-
-        super().__init__(
-            source=Relation.__coerce__(source),
-            selector=hashable_selector,
-            by=by,
-            sample=sample,
-        )
-
-    def __repr__(self) -> str:
-        return _semantic_repr(self)
-
-    @property
-    def values(self) -> FrozenOrderedDict[str, Any]:
-        from .._xorq import api as xo
-
-        return FrozenOrderedDict(
-            {
-                "fieldName": xo.literal("").op(),
-                "fieldPath": xo.literal("").op(),
-                "fieldType": xo.literal("").op(),
-                "fieldValue": xo.literal("").op(),
-                "weight": xo.literal(0).op(),
-            },
-        )
-
-    @property
-    def schema(self) -> Schema:
-        return Schema(
-            {
-                "fieldName": "string",
-                "fieldPath": "string",
-                "fieldType": "string",
-                "fieldValue": "string",
-                "weight": "int64",
-            },
-        )
-
-    @property
-    def keys(self) -> tuple[str, ...]:
-        return ("fieldValue", "fieldName", "fieldPath", "fieldType")
-
-    @property
-    def aggs(self) -> dict[str, Any]:
-        return {"weight": lambda t: t.weight}
-
-    def to_untagged(self):
-        all_roots = _find_all_root_models(self.source)
-        base_tbl = (
-            _to_untagged(self.source).limit(self.sample)
-            if self.sample
-            else _to_untagged(self.source)
-        )
-
-        merged_dimensions = _get_merged_fields(all_roots, "dimensions")
-        fields_to_index = _get_fields_to_index(
-            self.selector,
-            merged_dimensions,
-            base_tbl,
-        )
-
-        if not fields_to_index:
-            from .._xorq import api as xo
-
-            return xo.memtable(
-                {
-                    "fieldName": [],
-                    "fieldPath": [],
-                    "fieldType": [],
-                    "fieldValue": [],
-                    "weight": [],
-                },
-            )
-
-        def build_fragment(field_name: str) -> Any:
-            field_expr = (
-                merged_dimensions[field_name](base_tbl)
-                if field_name in merged_dimensions
-                else base_tbl[field_name]
-            )
-            field_type = field_expr.type()
-            type_str = _get_field_type_str(field_type)
-            weight_expr = _get_weight_expr(
-                base_tbl,
-                self.by,
-                all_roots,
-                field_type.is_string(),
-            )
-
-            return (
-                _build_string_index_fragment(
-                    base_tbl,
-                    field_expr,
-                    field_name,
-                    field_name,
-                    type_str,
-                    weight_expr,
-                )
-                if field_type.is_string() or not field_type.is_numeric()
-                else _build_numeric_index_fragment(
-                    base_tbl,
-                    field_expr,
-                    field_name,
-                    field_name,
-                    type_str,
-                    weight_expr,
-                )
-            )
-
-        fragments = [build_fragment(f) for f in fields_to_index]
-        return reduce(lambda acc, frag: acc.union(frag), fragments[1:], fragments[0])
-
-    def filter(self, predicate: Callable) -> SemanticFilter:
-        from ..expr import SemanticFilter
-
-        return SemanticFilter(source=self, predicate=predicate)
-
-    def order_by(self, *keys: str | ir.Value | Callable) -> SemanticOrderBy:
-        from ..expr import SemanticOrderBy
-
-        return SemanticOrderBy(source=self, keys=keys)
-
-    def limit(self, n: int, offset: int = 0) -> SemanticLimit:
-        from ..expr import SemanticLimit
-
-        return SemanticLimit(source=self, n=n, offset=offset)
-
-    def execute(self):
-        return _rebind_to_canonical_backend(self.to_untagged()).execute()
-
-    def as_expr(self):
-        """Return self as expression."""
-        return self
-
-    def compile(self, **kwargs):
-        return self.to_untagged().compile(**kwargs)
-
-    def sql(self, **kwargs):
-        return ibis.to_sql(self.to_untagged(), **kwargs)
-
-    def __getitem__(self, key):
-        return self.to_untagged()[key]
-
-    def pipe(self, func, *args, **kwargs):
-        return func(self, *args, **kwargs)
+from ._index_op import (  # noqa: E402
+    SemanticIndexOp,
+    _build_numeric_index_fragment,
+    _build_string_index_fragment,
+    _get_field_type_str,
+    _get_fields_to_index,
+    _get_weight_expr,
+    _resolve_selector,
+)
 
 
 from ._root_models import (  # noqa: E402
